rename signal submodule to audio
