+++ pyfar/classes/audio.py
